box: store request type in 16-bit interger
It seems that it'll never need it greater that UINT16_MAX.
Maybe it will allow us to save some memory.

Part of #5958

(cherry picked from commit 6e6554cb0fe29ee04a7141c5e43ed428a8bebd09)

+++ src/box/memtx_space.c
@@ -315,7 +315,7 @@ rollback:
 }
 
 static inline enum dup_replace_mode
-dup_replace_mode(uint32_t op)
+dup_replace_mode(uint16_t op)
 {
 	return op == IPROTO_INSERT ? DUP_INSERT : DUP_REPLACE_OR_INSERT;
 }
